+++ js/utils.js
@@ -1,6 +1,3 @@
-import { NAMES, COMMENTS } from './data.js';
-
-
 // Функция случайного числа из выбрранного диапозона -
 const getRandomInt = function (from, to) {
   if (to > from) {
@@ -21,25 +18,6 @@ const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.le
 // Функция проверки длины строки
 const checkingStringLength = (string, maxStringLength) => string.length <= maxStringLength;
 
-//Функция случайного 1-2 комментариев -
-const getTextComments = () => {
-  let text = '';
-  for (let i = 1; i <= getRandomInt(1, 2); i++) {
-    text += getRandomArrayElement(COMMENTS);
-  }
-  return text;
-};
-
-//Функция генерации комментария -
-const getComments = (j) => {
-  j++;
-  return {
-    id: j,
-    avatar: `img/avatar-${getRandomInt(1, 6)}.svg`,
-    message: getTextComments(),
-    name: getRandomArrayElement(NAMES),
-  };
-};
 
 //Нажатие ESC -
 const isEscapeKey = (evt) => evt.key === 'Escape';
@@ -47,4 +25,4 @@ const isEscapeKey = (evt) => evt.key === 'Escape';
 //Нажатие Enter
 const isEnterKey = (evt) => evt.key === 'Enter';
 
-export { getComments, checkingStringLength, getRandomArrayElement, getRandomInt, isEscapeKey, isEnterKey };
+export { checkingStringLength, getRandomArrayElement, getRandomInt, isEscapeKey, isEnterKey };
